Use uint16_t for storing width & height of the rendered image

+++ include/reelmagic.h
@@ -39,7 +39,7 @@ struct ReelMagic_VideoMixerMPEGProvider {
 
 void ReelMagic_RENDER_SetPal(uint8_t entry, uint8_t red, uint8_t green, uint8_t blue);
 
-void ReelMagic_RENDER_SetSize(const uint32_t width, const uint32_t height,
+void ReelMagic_RENDER_SetSize(const uint16_t width, const uint16_t height,
                               const bool double_width, const bool double_height,
                               const Fraction& render_pixel_aspect_ratio,
                               const uint8_t bits_per_pixel,
